Fix: allow saving new objects after deleting field values (parse-community#1751)
